refactor(client): reorganize auth feature directory
- Move auth pages into subdirectories within `features/auth/pages`
- Move auth components into `features/auth/components`
- Update exports in `features/auth/components/index.ts` and `features/auth/pages/index.ts`
