--- apps/client/src/features/auth/components/buttons/index.ts
@@ -1,3 +1,0 @@
-export * from "./auth-submit-button";
-export * from "./google-auth-button";
-export * from "./logout-button";
