--- apps/client/src/features/auth/components/forms/index.ts
@@ -1,4 +1,0 @@
-export * from "./login-form";
-export * from "./register-form";
-export * from "./forgot-password-form";
-export * from "./reset-password-form";
