+++ apps/client/src/features/auth/components/index.ts
@@ -1,6 +1,10 @@
 export * from "./auth-initializer";
-export * from "./buttons";
-export * from "./forms";
-
-export * from "./links";
-export * from "./ui";
+export * from "./auth-page-layout";
+export * from "./auth-submit-button";
+export * from "./email-confirmation-alert";
+export * from "./forgot-password-link";
+export * from "./google-auth-button";
+export * from "./logout-button";
+export * from "./password-reset-error-state";
+export * from "./password-reset-loading-state";
+export * from "./social-login-options";
