--- apps/client/src/features/auth/components/links/index.ts
@@ -1,1 +1,0 @@
-export * from "./forgot-password-link";
