+++ apps/client/src/features/auth/components/password-reset-error-state.tsx
