+++ apps/client/src/features/auth/components/social-login-options.tsx
@@ -1,4 +1,4 @@
-import { GoogleAuthButton } from "../buttons";
+import { GoogleAuthButton } from "./";
 
 export function SocialLoginOptions() {
   return (
