--- apps/client/src/features/auth/components/ui/index.ts
@@ -1,5 +1,0 @@
-export * from "./auth-page-layout";
-export * from "./email-confirmation-alert";
-export * from "./password-reset-error-state";
-export * from "./password-reset-loading-state";
-export * from "./social-login-options";
