+++ apps/client/src/features/auth/pages/forgot-password/components/index.ts
@@ -1,0 +1,1 @@
+export * from "./forgot-password-form";
