+++ apps/client/src/features/auth/pages/index.ts
@@ -1,4 +1,4 @@
-export * from "./login-page";
-export * from "./register-page";
-export * from "./reset-password-page";
-export * from "./forgot-password-page";
+export * from "./forgot-password";
+export * from "./login";
+export * from "./register";
+export * from "./reset-password";
