+++ apps/client/src/features/auth/pages/register/components/index.ts
@@ -1,0 +1,1 @@
+export * from "./register-form";
