+++ apps/client/src/features/auth/pages/register/components/register-form.tsx
@@ -9,9 +9,8 @@ import { useForm } from "react-hook-form";
 
 import { FormInput } from "@/components/global";
 import { Form } from "@/components/ui";
-
-import { useAuth } from "../../hooks";
-import { AuthSubmitButton } from "../buttons";
+import { AuthSubmitButton } from "@/features/auth/components";
+import { useAuth } from "@/features/auth/hooks";
 
 export function RegisterForm() {
   const form = useForm<IRegisterDto>({
