+++ apps/client/src/features/auth/pages/reset-password/components/index.ts
@@ -1,0 +1,1 @@
+export * from "./reset-password-form";
