+++ apps/client/src/features/auth/pages/reset-password/index.ts
@@ -1,0 +1,1 @@
+export * from "./page";
